requirements.txt
web3>=6.0.0
python-telegram-bot>=20.0
httpx>=0.27.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1
psycopg_pool>=3.1
